resolve #11 Refactor generate rescue time data module
