+++ src/rescuetime/index.ts
@@ -1,48 +1,21 @@
-import { generateUrl, requestAll } from '../utils/api'
-import { getQueryString } from '../utils/parseUtils'
+import { requestAll } from '../utils/api'
 import { getDayilDataSummary, compareWithYesterday } from './dailyReportUtils'
 import {
   parseOverview,
   parseActivity,
   parseEfficiency,
-  getWeekRange,
 } from './weeklyReportUtils'
-import { getToday } from '../utils/misc'
+import { dailyDataRequestUrls, weeklyDataRequestUrls } from '../utils/urls'
 
 export const getDailyData = async () => {
-  const t = getToday()
-  const today = t.subtract(1, 'day').format('YYYY-MM-DD')
-  const yesterday = t.subtract(2, 'day').format('YYYY-MM-DD')
-
-  const queryObject = {
-    format: 'json',
-    rb: today,
-    re: today,
-    rk: 'efficiency',
-  }
-
-  const todayEfficiencyUrl = `${generateUrl('/data')}&${getQueryString(queryObject)}`
-  const yesterDayEfficiencyUrl = `${generateUrl('/data')}&${getQueryString({
-    ...queryObject,
-    rb: yesterday,
-    re: yesterday,
-  })}`
-
-  const todayOverviewUrl = `${generateUrl('/data')}&${getQueryString({
-    ...queryObject,
-    rk: 'overview',
-  })}`
-  const yesterdayOverviewUrl = `${generateUrl('/data')}&${getQueryString({
-    ...queryObject,
-    rb: yesterday,
-    re: yesterday,
-    rk: 'overview',
-  })}`
-
-  const activityUrl = `${generateUrl('/data')}&${getQueryString({
-    ...queryObject,
-    rk: 'activity',
-  })}`
+  const {
+    todayEfficiencyUrl,
+    yesterDayEfficiencyUrl,
+    todayOverviewUrl,
+    yesterdayOverviewUrl,
+    activityUrl,
+    today,
+  } = dailyDataRequestUrls
 
   try {
     const [todayEfficiency, yesterDayEfficiency, todayOverview, yesterdayOverview, activity] = await requestAll([
@@ -76,26 +49,13 @@ export const getDailyData = async () => {
 
 export const getWeeklyData = async () => {
   try {
-    const { from, to } = getWeekRange()
-
-    const queryObject = {
-      format: 'json',
-      rb: from,
-      re: to,
-      rk: 'overview',
-    }
-
-    const overviewUrl = `${generateUrl('/data')}&${getQueryString(
-      queryObject,
-    )}`
-    const activityUrl = `${generateUrl('/data')}&${getQueryString({
-      ...queryObject,
-      rk: 'activity',
-    })}`
-    const efficiencyUrl = `${generateUrl('/data')}&${getQueryString({
-      ...queryObject,
-      rk: 'efficiency',
-    })}`
+    const {
+      overviewUrl,
+      activityUrl,
+      efficiencyUrl,
+      from,
+      to,
+    } = weeklyDataRequestUrls
 
     const [overview, activity, efficiency] = await requestAll([
       overviewUrl,
