+++ src/rescuetime/weeklyReportUtils.ts
@@ -1,22 +1,10 @@
 import { map } from 'ramda'
 
-import { getToday } from '../utils/misc'
 import { Overview, Activity, Efficiency } from '../@types/models'
 
 export type ParsedOverview = ReturnType<typeof parseOverview>[0]
 export type ParsedActivity = ReturnType<typeof parseActivity>[0]
 export type Parsedefficiency = ReturnType<typeof parseEfficiency>[0]
-
-export const getWeekRange = () => {
-  const currentDate = getToday()
-  const from = currentDate.subtract(6, 'day').format('YYYY-MM-DD')
-  const to = currentDate.subtract(1, 'day').format('YYYY-MM-DD')
-
-  return {
-    from,
-    to,
-  }
-}
 
 export const parseOverview = (overview: Overview) =>
   map(
