+++ src/utils/misc.ts
@@ -9,6 +9,17 @@ import { Parsedefficiency } from '../rescuetime/weeklyReportUtils'
 
 export const getSpentTime = (ce?: Parsedefficiency) => ce?.timeSpent ?? 0
 
+export const getWeekRange = () => {
+  const currentDate = getToday()
+  const from = currentDate.subtract(6, 'day').format('YYYY-MM-DD')
+  const to = currentDate.subtract(1, 'day').format('YYYY-MM-DD')
+
+  return {
+    from,
+    to,
+  }
+}
+
 export const mergeTimeByCategory = (efficiencies: Parsedefficiency[]) => {
   const ce = camelizeKeys(indexBy(prop('efficiency'), efficiencies)) as {[key: string]: Parsedefficiency}
